@shogas 2018-11-21
It should also be noted that passing `parallel=False` to `map` helps a lot on my machine. The inner loop is quite tight, and the global interpreter lock only allows one thread running python code at the same time. It might be interesting to try to write the inner loop as a numpy computation directly (multiple threads can run numpy at the same time), but I don't immidiately see how.