@dnjohnstone 2018-11-21
Thanks @shogas!

@pc494 this looks good to me - can you confirm you're happy with it before I merge.